# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs for Shop Scrape App",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pack-sizes": {
            "post": {
                "tags": [
                    "Pack Sizes"
                ],
                "summary": "Retrieve pack sizes",
                "description": "Fetches a list of pack sizes with optional pagination.",
                "operationId": "a71003c272c1c9e5b1de18ca5a083dc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dataPerPage": {
                                        "description": "Number of items per page",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "page": {
                                        "description": "Page number for pagination",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PackSizeResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/pack-sizes/store": {
            "post": {
                "tags": [
                    "Pack Sizes"
                ],
                "summary": "Store a pack size",
                "description": "Creates a new pack size.",
                "operationId": "e0042a376b94a2309dcd3764cc919d1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the pack size",
                                        "type": "string",
                                        "example": "Small"
                                    },
                                    "description": {
                                        "description": "Description of the pack size",
                                        "type": "string",
                                        "example": "Small sized package"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pack size created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PackSizeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pack-sizes/{packSize}": {
            "put": {
                "tags": [
                    "Pack Sizes"
                ],
                "summary": "Update a pack size",
                "description": "Updates an existing pack size.",
                "operationId": "f02d341ed512536182abe913c4c5e924",
                "parameters": [
                    {
                        "name": "packSize",
                        "in": "path",
                        "description": "ID of the pack size to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the pack size",
                                        "type": "string",
                                        "example": "Medium"
                                    },
                                    "description": {
                                        "description": "Updated description of the pack size",
                                        "type": "string",
                                        "example": "Medium sized package"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pack size updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PackSizeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pack Sizes"
                ],
                "summary": "Delete a pack size",
                "description": "Deletes a pack size by ID.",
                "operationId": "02d0ef2987cc136ed494ae7a37a8b384",
                "parameters": [
                    {
                        "name": "packSize",
                        "in": "path",
                        "description": "ID of the pack size to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pack size deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Operation success status",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string",
                                            "example": "Pack size deleted successfully."
                                        },
                                        "data": {
                                            "description": "Additional response data (empty for delete operation)",
                                            "type": "object",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PackSizeResource": {
                "description": "Pack Size resource schema",
                "properties": {
                    "id": {
                        "description": "ID of the pack size",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the pack size",
                        "type": "string",
                        "example": "Small"
                    },
                    "weight": {
                        "description": "Weight of the pack size with unit",
                        "type": "string",
                        "example": "500 g"
                    },
                    "amount": {
                        "description": "Amount of items in the pack",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}